move type tests to own file
